move staticFixtures to separate file.
