+++ vc-generator/helpers.js
@@ -1,9 +1,6 @@
 /*!
  * Copyright 2023-2024 Digital Bazaar, Inc.
  */
-import {createRequire} from 'node:module';
-
-const require = createRequire(import.meta.url);
 
 export function invalidCreateProof({
   addCreated = true,
@@ -98,14 +95,4 @@ export function invalidCreateProof({
 
     return proof;
   };
-}
-
-export function getStaticFile({
-  suiteName,
-  fileName,
-  version
-}) {
-  const filePath = `../inputs/${suiteName}/${version}-${fileName}` +
-    `-SimpleSigned2.json`;
-  return require(filePath);
 }
